+++ src/main/java/com/layoutmanager/layout/store/create/NewLayoutAction.java
@@ -11,7 +11,7 @@ import com.layoutmanager.layout.store.LayoutCreator;
 import com.layoutmanager.localization.MessagesHelper;
 import com.layoutmanager.persistence.Layout;
 import com.layoutmanager.persistence.LayoutConfig;
-import com.layoutmanager.ui.helpers.NotificationHelper;
+import com.layoutmanager.ui.helpers.BaloonNotificationHelper;
 import com.layoutmanager.ui.menu.WindowMenuService;
 
 import org.jetbrains.annotations.NotNull;
@@ -49,7 +49,7 @@ public class NewLayoutAction extends AnAction {
     }
 
     private void showNotification(Layout newLayout) {
-        NotificationHelper.info(
+        BaloonNotificationHelper.info(
                 MessagesHelper.message("StoreLayout.New.Notification.Title"),
                 MessagesHelper.message("StoreLayout.New.Notification.Content", newLayout.getName()));
     }
