+++ src/main/resources/com/layoutmanager/ui/messages.properties
@@ -28,3 +28,7 @@ SettingsPage.NameColumn=Name
 SettingsPage.ConfiguredWindowsColumn=Configured Windows
 
 ImportDialog.SelectFileTitle=Select layout file
+ImportDialog.UnexpectedError=Unexpected error occurred: {0}
+ImportDialog.SuccessfullyLoadedLayout=Successfully loaded the layout ''{0}''
+
+LayoutNameValidation.InvalidName=The layout name must not be empty or contain illegal characters.
